many changes has been made in this commit: 1. integrate glm math lib 2. implement camera 3. refactor the framework 4. implement render component 5. implement point cloud rendering

+++ XRealityRender/src/XREntity.cpp
@@ -5,7 +5,7 @@ bool XREntity::init()
 	return true;
 }
 
-bool XREntity::update(float time)
+bool XREntity::update(double time)
 {
 	return true;
 }
