+++ XRealityRender/src/XREntity.h
@@ -3,11 +3,17 @@
 
 #include "XREngine.h"
 
+/**
+ * XREntity
+ * XREntity is an game entity in a general sense. It can be GameObject, UI, effects, etc.
+ * 
+ * @Author Yu Mao
+ */
 class XREntity
 {
 public:
 	virtual bool init();
-	virtual bool update(float time);
+	virtual bool update(double time);
 	virtual bool destroy();
 };
